Clippy: clean up

+++ tests/assignment_expression_test.rs
@@ -102,5 +102,5 @@ fn plus_equal_num_addition() {
     // When
     Vm::execute_program(vm, function);
     // Then
-    assert_eq!(3 as u32, events.borrow().get("a").unwrap().value.number_value());
+    assert_eq!(3_u32, events.borrow().get("a").unwrap().value.number_value());
 }
